Add method for respawn mobs

+++ core/src/com/parabells/PTCommands/MovM.java
@@ -1,11 +1,11 @@
 package com.parabells.PTCommands;
 
 public class MovM extends DefaultCommand {
-    public int x;
-    public int y;
+    public float x;
+    public float y;
     public int[] mobs;
 
-    public MovM(int x, int y, int[] mobs){
+    public MovM(float x, float y, int[] mobs){
         this.x = x;
         this.y = y;
         this.mobs = mobs;
